requirements.txt
django
djangorestframework
boto3
awscli
django-redis